Refactoring gm, new soundclass
Refactoring GameplayManager ->
Moved all projectile creation and gun update calls to the ship class (I
feel like the ship should know if it can shot, if so create the
projectile).
Added a get function for the projectileList
Added a clear function for the projectile list, which are both called
from the gm(*collision checks, might throw all collision checks into a
seperate class in future).
Added A static SoundClass, because of easier access, maintainability and
easier implementation of new soundeffects.
Removed unnecessary variables associated with explosion timers and the
timers themselves.(honestly do not know why they were there to begin
with..)