Further Refactoring of GameManager, new Collision Manager, new EffectsManager
GameManager ->
Pulled all collsion logic into a seperate(helper) class. It should now
be much more readable and easier to maintain. This class is created by
the gm.
Pulled all effect logics into a seperate class. Same reasons as above.
This is a static class, might remake it into a normal one, but seeing
that the afterburner effect for the ship gets called from the ship
itself, to keep it easy to implement new effects I'm keeping it static
for now.
Removed unnecessary variables.