Tweaked Afterburner EmitterLocation, further experimentation with colourpallets.
Afterburner Emitter location should now Emit it's particles from the
rear of the ship.
Afterburner now uses a colourpallet from which it selects a colour.
Reintroduced the "hard" particlesPerTick variable to the particleEngine.
ParticleEngine now uses an intensifier variable derived from the ship
speed as a multiplier to the particlesPerTick variable.

Next up is an attempt to jetstream the afterburner, to mkae it look more
like an afterburner.