tweaks to missiles
tweaked how the rotation works with the missiles, they should no long
wobble, they now have a nice smooth detach from the ship followed by an
acceleration towards the targetposition(actually a bit infront of the
targetposition so that the player doesn't have to aim at the side of the
screen to hit something there, which would have caused the ship to
accelerate towards whatever was being aimed at)
More tweaks and probably some balancing are going to be done to the
missiles as noted at the previous push. Ex, they need a texture(not just
a grey blub)  I also want them to have an afterburner, though I am
considering moving away from the particle engine in favor for
spritesheets, or at the very least switch the textures used by the
particle engine ultimately creating 100-1000 particles isn't great I
knew this going in, I'll have some more coffee and pounder for a bit.
Same thing goes for the asteroid explosions ship afterburner.