miniMissileTex added
v1 of missile tex. Not a big push ..

Also do not think I haven't noticed that the bezier curve is..
uneccesary, creating alot of vectors that are never used, in reality,
for the detach visual I've been going for makes the original bezier
curve path obsolete. It will be removed/reworked and probably saved for
future projectiles.