Added a proper healthbar,  experimenting with the UI design,. Experimented with new particles for p-engine
healthbar ->
Added texture, and funtionality for the healthbar, works as intended but
the design will have to be developed further. I have a vague idea for
the design in general. The current projectory is leaning more towards
proper models as opposet to pixel graphics. With that in mind I would
see the UI and HUD of a similar design, as for the layout, I'd like to
keep it minimalistic, but always provide the player with information. An
example might be to make the healthbar small and have the hover over the
ship and enemy ships,  alongside ammo count, shield capacity so on so
forth whilst keeping other ui elemtns towards the left/bottom of the
screen.

particles for p-engine ->
Added new texture for the particle engine. Tweaked values for
asteroidexplosion-effect. It now creates fewer but bigger
(multicoloured) particles. With the right tweaks this might work out.
I'll make a spritesheet and do some experimentation with that before I
decide.
As I mentioned in a previous push, I've been thinking about moving away
from the p-engine, mainly because as the game grows the more effects may
happen on screen at the same time, to save space and make the game
handle (and continue to handle as it grows) the increasing load, it
might be prudent to use spritesheet animations for explosions and other
effects. Example might be a spritesheet animation for asteroid
explosions and on-hit explosions. But might keeping the particle
generator for pick-up items and other sparser effects...