New Entity (Drone) added
Drone ->
player (and enemy) support entity. Still in a basic form it follows the
playership in a circular pattern, engaging asteroids as they come to
close.

There are some obvious issues, for one the little devil teleports back
to its circular path after engage, I would like to see it fly over there
instead. Changes need to be made to the laser canon (and all  gun
classes) as so they dont all fire on mouse input from the player... in
the case of the drone, is it a bug or a feature... Still, changes needs
to be made. Going to reduce the size and some balancing once I get the
immidiate bugs out the way. Also a Chase (block asteroid from hitting
ship) and an Evade state will be added to the Drone