Fixes to Drone
Shouldn't crash anymore. But I will have to take another pass at how the
drone works and another pass how the data is sent around.
Two main issues with the Drone as I've noted are 1: Sometimes an
explosion is not triggered.
2: Sometimes the Drone gets stuck in a loop engaging a target that is
not there anymore.

Also moved the left click to shoot logic into the ship class, will make
it a fire primary and secondary, these functions will most likely be
stored in the ship base class and overloaded for the player, but any ai
controlled ship could use the functions(and so could the drone)

Fixes to Drone ->
"Fixed" the EngageTarget function as so it will no longer compare values
greater than the array size, it also knows if the target has been
removed. Some issues remain 2 of which is mentioned above, also im not
happy about the code.